// src/lib/flex.types.ts
export type T_FlexDirection =
  | "flex-row"
  | "flex-row-reverse"
  | "flex-col"
  | "flex-col-reverse";

export type T_JustifyContent =
  | "justify-start"
  | "justify-end"
  | "justify-center"
  | "justify-between"
  | "justify-around"
  | "justify-evenly";

export type T_AlignItems =
  | "items-stretch"
  | "items-center"
  | "items-start"
  | "items-end";
  
export type T_AlignSelf =
  | "self-auto"
  | "self-start"
  | "self-end"
  | "self-center"
  | "self-stretch";
  
export type T_FlexWrap = "flex-nowrap" | "flex-wrap" | "flex-wrap-reverse";